# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API v1.0.0",
        "contact": {
            "email": "prasdimasetyo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/skills": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get a list of all skills",
                "operationId": "6acb38b223d6471cc32eab45ea57be4b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "skills": [
                                        {
                                            "id": 1,
                                            "name": "Programming",
                                            "created_by": null,
                                            "created_at": "2024-02-20T14:52:29.000000Z",
                                            "updated_by": null,
                                            "updated_at": "2024-02-20T14:52:29.000000Z",
                                            "deleted_by": null,
                                            "deleted_at": null,
                                            "skillSets": []
                                        },
                                        {
                                            "id": 2,
                                            "name": "Design",
                                            "created_by": null,
                                            "created_at": "2024-02-20T14:52:29.000000Z",
                                            "updated_by": null,
                                            "updated_at": "2024-02-20T14:52:29.000000Z",
                                            "deleted_by": null,
                                            "deleted_at": null,
                                            "skillSets": []
                                        },
                                        {
                                            "id": 3,
                                            "name": "Testing",
                                            "created_by": null,
                                            "created_at": "2024-02-20T14:52:29.000000Z",
                                            "updated_by": null,
                                            "updated_at": "2024-02-20T14:52:29.000000Z",
                                            "deleted_by": null,
                                            "deleted_at": null,
                                            "skillSets": []
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skills not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skills not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Skills"
                ],
                "summary": "Create a new skill",
                "operationId": "4725cd560d3046ae33bd690208e21d41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Skill"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skill created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "skill": {
                                        "id": 4,
                                        "name": "New Skill",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "skillSets": []
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/skills/{id}": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "Get a skill by ID",
                "operationId": "81b9cbc27bbdd8b5a6591975bf75441a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "skill": {
                                        "id": 1,
                                        "name": "Programming",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "skillSets": []
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skill not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Skills"
                ],
                "summary": "Update a skill by ID",
                "operationId": "6c932a7a6f40be346ab93e4dd28f14f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Skill"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "skill": {
                                        "id": 4,
                                        "name": "Updated Skill",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "skillSets": []
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skill not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Skills"
                ],
                "summary": "Delete a skill by ID",
                "operationId": "7401a82a4b0de61e80c67901e89d62dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Skill ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Skill deleted successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skill not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Skill not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}